+++ lib/oxidized/model/apc_aos.rb
@@ -2,9 +2,10 @@ class Apc_aos < Oxidized::Model # rubocop:disable Naming/ClassAndModuleCamelCase
   using Refinements
 
   cmd 'config.ini' do |cfg|
-    cfg.gsub! /^; Configuration file, generated on.*/, ''
+    cfg.gsub!(/^; Configuration file, generated on.*\n/, '')
+    cfg
   end
 
-  cfg :ftp do
+  cfg :ftp, :scp do
   end
 end